ApiImplementationWithoutJupyter/requirements.txt
# jupyter-client==6.1.11
# jupyter-core==4.7.0
# jupyter-server==1.13.5
# jupyter-server-mathjax==0.2.2
# jupyterlab==3.3.1
# jupyterlab-pygments==0.1.2
# jupyterlab-server==2.10.3
# notebook==6.4.8
PyYAML==6.0
Pillow>=7.1.2
requests==2.27.1
requests-oauthlib==1.3.0
pydantic==1.7.4
python-multipart==0.0.5
uvicorn==0.17.6
fastapi==0.75.0
tqdm==4.63.1

seaborn>=0.11.0
matplotlib==3.5.1
scipy==1.8.0
numpy==1.22.3
plotly==4.14.3
pandas==1.4.1
# opencv-python==4.5.5.64

# scikit-learn==0.24.0
# torch==1.11.0
# torchvision==0.12.0


dlib==19.23.1
opencv-contrib-python==4.5.5.64
torch==1.11.0
torchvision==0.12.0